checkpoint integration

+++ admin.sh
@@ -33,6 +33,13 @@ then
   exit
 fi
 
+
+if [[ -z "${PROJECT_ROOT}" ]];
+then
+  echo "Error: environment variable PROJECT_ROOT not set."
+  exit
+fi
+
 test -d test_sgx || mkdir test_sgx
 cd ./test_sgx
 rm -rf *
@@ -67,6 +74,8 @@ fi
 ### 0.2 BulletinBoard INIT: ... ###
 ../stop_BB.sh
 ../run_BB.sh
+export BILLBOARD_URL="http://$(docker inspect -f '{{range.NetworkSettings.Networks}}{{.IPAddress}}{{end}}' billboard):8545"
+echo "BILLBOARD_URL=$BILLBOARD_URL"
 python3 ../gen_user_keys.py
 
 
@@ -90,22 +99,12 @@ cp sealedState.bin sealedState0.bin
 echo ""
 echo "[biPVRA] Running Auditee to Extract PVRA signing_key"
 
-source ~/.venvs/auditee/bin/activate
-python3.7 ../auditee_extract.py
+#source ~/.venvs/auditee/bin/activate
+#python3.7 ../auditee_extract.py
+python3 ../auditee_extract.py
+#source ~/.venvs/auditee/bin/activate deactivate
 
-
-
-
-# [TODO][NERLA]: ias_report.json needs to be posted to Bulletin Board
-# Let me know if it needs to be signed by the enclave
-# This is returned AFTER initPVRA returns the quote, quote is sent to intel and ias_report.json comes back
-# We will have to run ANOTHER ecall to sign it by the enclave
-# Alternatively the admin has a key pair, and the admin signs it because the admin is the person with the SPID that requests the ias_report from intel?
-
-
-
-
-
+# [TODO][NERLA]: ask sylvain which part of ias_report to post in order to verify enclave sig offline
 
 ### 2.0 PREPARE TO RUN PVRA APPLICATION ###
 
@@ -114,32 +113,8 @@ python3.7 ../auditee_extract.py
 echo -n "[biPVRA] Verifying signed encryption key: "
 openssl dgst -sha256 -verify signingkey.pem -signature enclave_enc_pubkey.sig enclave_enc_pubkey.bin
 
-
-
-
-
-### 2.2 Verify signed user pubkeys using signingkey.pem (extracted enclave signing key), the same action carried out by the billboard ###
-
-# The blob enclave signed: hash(64b[user0_pubkey] || 64b[user1_pubkey] || ... || 64b[user4_pubkey])
-# This bash code recreates that blob and hash
-
-touch pubkeys_to_be_hashed.bin
-
-line=$(head -n 1 pubkeys.list)
-num_keys=$(($line))
-iter=2
-
-while [ $num_keys != 0 ]
-do
-  out=$(awk "FNR>=${iter} && FNR <=${iter}" pubkeys.list | xxd -p -r >> pubkeys_to_be_hashed.bin)
-  num_keys=$(($num_keys-1))
-  iter=$(($iter+1))
-done
-
-echo -n "[biPVRA] Verifying signed userpubkeys: "
-openssl dgst -sha256 -verify signingkey.pem -signature pubkeys.sig pubkeys_to_be_hashed.bin
-
-
+echo -n "[biPVRA] Initialize billboard and Verifying signed userpubkeys:"
+python3 $PROJECT_ROOT/billboard/billboard.py admin_init_contract pubkeys.list pubkeys.sig
 
 
 ### 2.2 SETUP CLIENT ENVIRONMENT ###
@@ -156,10 +131,10 @@ cp ../format_command ./client
 cp ./user0_pubkey.bin ./client
 cp ./user0_prikey.bin ./client
 
-# [TODO][NERLA]: If you write your ECDH script in client_ecdh.py nothing needs to be done
 # Otherwise copy your script to client environment
 cp ../client_ecdh.py ./client
 
+cp ../gen_ecdh.py ./client
 
 ### 2.3 SETUP HOST ENVIRONMENT ###
 mkdir host
@@ -168,6 +143,7 @@ cp ../pvraHostCommand.sh ./host
 cp ../pvraAuditCommand.sh ./host
 cp ./sealedState0.bin ./host
 cp ./signingkey.pem ./host
+cp ./signingkey.bin ./host
 if [[ ${CCF_ENABLE} == "1" ]];
 then
   cp ../service_cert.pem ./host
